Adding a simple, naive dump function
With this commit a really basic dump function comes into the game.
The HclBase class can be used to indicate wether a class is a HCL-Stanza
or not and depending on this base class - as well as NAME and _hcl_label
variables - the Stanza is then created. Nested Stanzas are already possible
Also the gitignore is extended to pycharm specific stuff.